Authentication and Authorization of API requests using JWT

+++ controllers/api/v1/posts_api.js
@@ -28,19 +28,19 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
             
-            // req.flash('success', 'Post and associated comments deleted');
             return res.json(200, {
                 message: "Post and associated comments deleted successfully"
             });
-        // }else{
-        //     req.flash('error', 'You cannot delete this post');
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
     }catch(err){
         // req.flash('error', err);
         console.log("****Error:", err);
